typos

+++ app/models/data.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-from sqlalchemy import Column, ForeignKey, BigInteger, Integer, Float, String, DateTime
+from sqlalchemy import Column, ForeignKey, Integer, Integer, Float, String, DateTime
 from sqlalchemy import UniqueConstraint, PrimaryKeyConstraint, Table
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
@@ -112,8 +112,8 @@ class ViewUserSkills(Base):
 
 	user_id = Column(Integer)
 	user_skill_id = Column(Integer)
-	answer_count = Column(BigInteger)
-	total_score = Column(BigInteger)
+	answer_count = Column(Integer)
+	total_score = Column(Integer)
 	__table_args__ = (PrimaryKeyConstraint('user_id', 'user_skill_id', name='view_user_skills_pk'),)
 
 def get_sqlite3_session(path):
